comments

+++ app/processing/fourier_filers.py
@@ -65,12 +65,12 @@ class FourierFilters:
                     # if so fill in the circle values with the mask value
                     mask[i, j] = mask_value
 
-        filtered_dft = image_fourier * mask    # multiply the fft of the image by the mask
-        dft_inverse = np.fft.ifftshift(filtered_dft)   #shift the spectrum back to the original format
-        filtered_image = np.fft.ifft2(dft_inverse)  #apply inverse fft
+        filtered_dft = image_fourier * mask                  # multiply the fft of the image by the mask
+        dft_inverse = np.fft.ifftshift(filtered_dft)         #shift the spectrum back to the original format
+        filtered_image = np.fft.ifft2(dft_inverse)           #apply inverse fft
 
-        filtered_image = np.real(filtered_image)   #extract only the real values of the filtered image to avoid complex ones
-        filtered_image = np.clip(filtered_image, 0, 255)   #clip pixels to the [0-255] range
-        filtered_image = filtered_image.astype(np.uint8)   #cast pixel values to uint8 
+        filtered_image = np.real(filtered_image)             #extract only the real values of the filtered image to avoid complex ones
+        filtered_image = np.clip(filtered_image, 0, 255)     #clip pixels to the [0-255] range
+        filtered_image = filtered_image.astype(np.uint8)     #cast pixel values to uint8 
 
         return filtered_image
